+++ src/ByteCodeDebugger_Break.cpp
@@ -12,7 +12,7 @@ void ByteCodeDebugger::printBreakpoints(ByteCodeRunContext* context)
     }
 
     g_Log.setColor(LogColor::Gray);
-    for (int i = 0; i < context->debugBreakpoints.size(); i++)
+    for (size_t i = 0; i < context->debugBreakpoints.size(); i++)
     {
         const auto& bkp = context->debugBreakpoints[i];
         g_Log.print(Fmt("#%d: ", i + 1));
