+++ src/ByteCodeGenJob_Function.cpp
@@ -2071,7 +2071,7 @@ bool ByteCodeGenJob::emitFuncDeclParams(ByteCodeGenContext* context)
     }
 
     auto childSize = node->childs.size();
-    for (int i = 0; i < childSize; i++)
+    for (size_t i = 0; i < childSize; i++)
     {
         if ((i == childSize - 1) && funcNode->typeInfo->flags & (TYPEINFO_VARIADIC | TYPEINFO_TYPED_VARIADIC))
             break;
