+++ src/SaveGenJob.cpp
@@ -8,7 +8,7 @@
 bool SaveGenJob::flush(Module* module)
 {
     ScopedLock lk(module->mutexFlushGenFiles);
-    for (int idx = 0; idx < module->contentJobGeneratedFile.size(); idx++)
+    for (size_t idx = 0; idx < module->contentJobGeneratedFile.size(); idx++)
     {
         if (module->contentJobGeneratedFile[idx].empty())
             continue;
